feat: separate out image and image view from texture
