+++ src/vulkan_engine.h
@@ -5,7 +5,7 @@
 
 #include "util/no_copy_or_move.h"
 #include "vulkan_context.h"
-#include "vulkan/texture_image.h"
+#include "vulkan/texture.h"
 
 namespace levin
 {
@@ -14,7 +14,7 @@ namespace levin
     private:
         std::unique_ptr<VulkanContext> m_context;
 
-        std::unique_ptr<TextureImage> m_texture_image;
+        std::unique_ptr<Texture> m_texture_image;
 
         uint32_t m_current_frame = 0;
